+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- Use route_id in timetable_page_id to fix issue with groupTimetableInfoPages
+- Fix for maps with frequency based trips
+
+### Updated
+- trip-frequency styles
+- Improvement for map hiding when no geojson available
+
 ## [2.12.6] - 2026-03-22
 
 ### Changed